requirements.txt
altgraph==0.16.1
astroid==2.2.5
colorama==0.4.1
Django==2.2
django-crispy-forms==1.7.2
future==0.17.1
gunicorn==19.9.0
isort==4.3.13
lazy-object-proxy==1.3.1
macholib==1.11
mccabe==0.6.1
pefile==2018.8.8
Pillow==6.0.0
PyInstaller==3.4
pylint==2.3.1
pytz==2019.1
pywin32-ctypes==0.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.1
wrapt==1.11.1
